AI_Agent_Project/pyproject.toml
[project]
name = "ai-agent-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "flask>=3.1.3",
    "google-generativeai>=0.8.6",
    "openpyxl>=3.1.5",
    "pandas>=3.0.2",
    "pyodbc>=5.3.0",
    "streamlit>=1.56.0",
]
